chore: updated worker to use the SDK
