+++ package.json
@@ -39,7 +39,7 @@
     "@ubiquity-dao/ubiquibot-logger": "^1.3.1",
     "@ubiquity-os/plugin-sdk": "^2.0.1",
     "dotenv": "16.4.5",
-    "typebox-validators": "0.3.5"
+    "hono": "^4.6.16"
   },
   "devDependencies": {
     "@commitlint/cli": "19.3.0",
@@ -65,7 +65,6 @@
     "npm-run-all": "4.1.5",
     "prettier": "3.3.3",
     "ts-jest": "29.1.5",
-    "tsx": "4.15.6",
     "typescript": "5.6.2",
     "typescript-eslint": "8.8.1",
     "wrangler": "3.102.0"
