+++ src/types/env.ts
@@ -1,7 +1,6 @@
-import { Type as T } from "@sinclair/typebox";
-import { StaticDecode } from "@sinclair/typebox";
+import { StaticDecode, Type as T } from "@sinclair/typebox";
 import "dotenv/config";
-import { StandardValidator } from "typebox-validators";
+import { LOG_LEVEL } from "@ubiquity-dao/ubiquibot-logger";
 
 /**
  * Define sensitive environment variables here.
@@ -10,8 +9,9 @@ import { StandardValidator } from "typebox-validators";
  * taken from either `dev.vars` or repository secrets.
  * They are used with `process.env` but are type-safe.
  */
-export const envSchema = T.Object({});
-
-export const envValidator = new StandardValidator(envSchema);
+export const envSchema = T.Object({
+  LOG_LEVEL: T.Optional(T.Enum(LOG_LEVEL)),
+  KERNEL_PUBLIC_KEY: T.Optional(T.String()),
+});
 
 export type Env = StaticDecode<typeof envSchema>;
