+++ src/types/plugin-inputs.ts
@@ -1,15 +1,4 @@
-import { SupportedEvents, SupportedEventsU } from "./context";
 import { StaticDecode, Type as T } from "@sinclair/typebox";
-import { StandardValidator } from "typebox-validators";
-
-export interface PluginInputs<T extends SupportedEventsU = SupportedEventsU, TU extends SupportedEvents[T] = SupportedEvents[T]> {
-  stateId: string;
-  eventName: T;
-  eventPayload: TU["payload"];
-  settings: PluginSettings;
-  authToken: string;
-  ref: string;
-}
 
 /**
  * This should contain the properties of the bot config
@@ -25,7 +14,5 @@ export const pluginSettingsSchema = T.Object(
   },
   { default: { configurableResponse: "Hello, world!" } }
 );
-
-export const pluginSettingsValidator = new StandardValidator(pluginSettingsSchema);
 
 export type PluginSettings = StaticDecode<typeof pluginSettingsSchema>;
